Allow findById to discriminate between ingredient/id
