+++ src/user-class.js
@@ -109,8 +109,6 @@ class User {
   }
 }
 
-
-
 if (typeof module !== 'undefined') {
   module.exports = User;
 }
