+++ test/pantry-test.js
@@ -236,15 +236,15 @@ describe('Pantry', () => {
     expect(requiredIngredientsInPantry).to.deep.equal(expectedIngredients);
   });
 
-  it('can find an ingredient\'s name from its Id', () => {
-    let name = pantry.findIngredientName(16069);
-    expect(name).to.equal('legumes');
-  });
+  // it('can find an ingredient\'s name from its Id', () => {
+  //   let name = pantry.findIngredientName(16069);
+  //   expect(name).to.equal('legumes');
+  // });
 
-  it('won\'t search for IDs that aren\'t numbers', () => {
-    let name = pantry.findIngredientName('legumes');
-    expect(name).to.equal(undefined);
-  })
+  // it('won\'t search for IDs that aren\'t numbers', () => {
+  //   let name = pantry.findIngredientName('legumes');
+  //   expect(name).to.equal(undefined);
+  // })
 
   it('should know how many ingredients are needed to make a recipe', () => {
     let missingIngredients = pantry.showMissingIngredients(greenHam);
